+++ aspnet/tutorials/first-mvc-app/start-mvc/sample/src/MvcMovie/Models/MovieDate.cs
@@ -1,4 +1,4 @@
-﻿#define AddDate
+﻿//#define AddDate
 #if AddDate
 using System;
 using System.ComponentModel.DataAnnotations;
