+++ aspnet/tutorials/first-mvc-app/start-mvc/sample/src/MvcMovie/Models/SeedData.cs
@@ -1,4 +1,6 @@
-﻿using Microsoft.Extensions.DependencyInjection;
+﻿//#define First
+#if First
+using Microsoft.Extensions.DependencyInjection;
 using MvcMovie.Data;
 using System;
 using System.Linq;
@@ -58,3 +60,5 @@ namespace MvcMovie.Models
         }
     }
 }
+
+#endif
