+++ aspnet/tutorials/first-mvc-app/start-mvc/sample/src/MvcMovie/Views/Movies/Create.cshtml
@@ -40,6 +40,13 @@
             </div>
         </div>
         <div class="form-group">
+            <label asp-for="Rating" class="col-md-2 control-label"></label>
+            <div class="col-md-10">
+                <input asp-for="Rating" class="form-control" />
+                <span asp-validation-for="Rating" class="text-danger" />
+            </div>
+        </div>
+        <div class="form-group">
             <div class="col-md-offset-2 col-md-10">
                 <input type="submit" value="Create" class="btn btn-default" />
             </div>
