+++ aspnet/tutorials/first-mvc-app/start-mvc/sample/src/MvcMovie/Views/Movies/Details.cshtml
@@ -34,6 +34,9 @@
         <dd>
             @Html.DisplayFor(model => model.Title)
         </dd>
+        <dd>
+            @Html.DisplayFor(model => model.Rating)
+        </dd>
     </dl>
 </div>
 <div>
